+++ tests/integration/aws_interface/test_service_resource_resource.py
@@ -6,11 +6,12 @@ from moto import mock_ec2, mock_iam, mock_s3, mock_sts
 from cloudwanderer.aws_interface.boto3_loaders import ResourceMap
 from cloudwanderer.urn import URN
 
-from ...pytest_helpers import compare_dict_allow_any, create_iam_role, create_s3_buckets
+from ...pytest_helpers import compare_dict_allow_any, create_iam_role, create_s3_buckets, get_single_ec2_vpc
 
 
 @mock_ec2
-def test_raw_data(single_ec2_vpc):
+@mock_sts
+def test_raw_data(ec2_service):
     compare_dict_allow_any(
         {
             "CidrBlock": "172.31.0.0/16",
@@ -29,7 +30,7 @@ def test_raw_data(single_ec2_vpc):
             "Tags": [],
             "VpcId": ANY,
         },
-        single_ec2_vpc.meta.data,
+        get_single_ec2_vpc(ec2_service).meta.data,
         allow_partial_match_first=True,
     )
 
@@ -38,7 +39,8 @@ def test_raw_data(single_ec2_vpc):
 # show OwnerId or not depending on the version. Needs further investigation.
 # Worked around for now by omitting the key and using allow_partial_match=True
 @mock_ec2
-def test_normalized_raw_data(single_ec2_vpc):
+@mock_sts
+def test_normalized_raw_data(ec2_service):
     compare_dict_allow_any(
         {
             "CidrBlock": "172.31.0.0/16",
@@ -57,20 +59,25 @@ def test_normalized_raw_data(single_ec2_vpc):
             "Tags": [],
             "VpcId": ANY,
         },
-        single_ec2_vpc.normalized_raw_data,
+        get_single_ec2_vpc(ec2_service).normalized_raw_data,
         allow_partial_match_first=True,
     )
 
 
-def test_resource_type(single_ec2_vpc):
-    assert single_ec2_vpc.resource_type == "vpc"
+@mock_ec2
+@mock_sts
+def test_resource_type(ec2_service):
+    assert get_single_ec2_vpc(ec2_service).resource_type == "vpc"
 
 
-def test_get_region_regional_resources(single_ec2_vpc):
-    assert single_ec2_vpc.get_region() == "eu-west-2"
+@mock_ec2
+@mock_sts
+def test_get_region_regional_resources(ec2_service):
+    assert get_single_ec2_vpc(ec2_service).get_region() == "eu-west-2"
 
 
 @mock_s3
+@mock_sts
 def test_get_region_global_service_global_resources(s3_service):
     create_s3_buckets(regions=["us-east-1", "eu-west-2", "ap-east-1"])
 
@@ -80,24 +87,30 @@ def test_get_region_global_service_global_resources(s3_service):
 
 
 @mock_sts
-def test_get_account_id(single_ec2_vpc):
-    assert single_ec2_vpc.get_account_id() == "123456789012"
+@mock_ec2
+def test_get_account_id(ec2_service):
+    assert get_single_ec2_vpc(ec2_service).get_account_id() == "123456789012"
 
 
-def test_service_name(single_ec2_vpc):
-    assert single_ec2_vpc.service_name == "ec2"
+@mock_ec2
+@mock_sts
+def test_service_name(ec2_service):
+    assert get_single_ec2_vpc(ec2_service).service_name == "ec2"
 
 
-def test_resource_map(single_ec2_vpc):
-    assert isinstance(single_ec2_vpc.resource_map, ResourceMap)
+@mock_ec2
+@mock_sts
+def test_resource_map(ec2_service):
+    assert isinstance(get_single_ec2_vpc(ec2_service).resource_map, ResourceMap)
 
 
 @mock_sts
-def test_get_urn(single_ec2_vpc):
-    assert isinstance(single_ec2_vpc.get_urn(), URN)
-    assert str(single_ec2_vpc.get_urn()).startswith(
+@mock_ec2
+def test_get_urn(ec2_service):
+    assert isinstance(get_single_ec2_vpc(ec2_service).get_urn(), URN)
+    assert str(get_single_ec2_vpc(ec2_service).get_urn()).startswith(
         "urn:aws:123456789012:eu-west-2:ec2:vpc"
-    ), f"{single_ec2_vpc.get_urn()} does not match 'urn:aws:123456789012:eu-west-2:ec2:vpc'"
+    ), f"{get_single_ec2_vpc(ec2_service).get_urn()} does not match 'urn:aws:123456789012:eu-west-2:ec2:vpc'"
 
 
 def test_dependent_resource_types(single_iam_role):
@@ -111,6 +124,7 @@ def test_collection(iam_service):
     single_iam_role = list(iam_service.collection("role").all())[0]
     single_role_policy = list(single_iam_role.collection("role_policy").all())[0]
     single_role_policy.load()
+    single_role_policy.fetch_secondary_attributes()
 
     assert (
         str(single_role_policy.get_urn()) == "urn:aws:123456789012:us-east-1:iam:role_policy:test-role/test-role-policy"
@@ -125,8 +139,10 @@ def test_collection(iam_service):
     }
 
 
-def test_secondary_attribute_names(single_ec2_vpc):
-    assert single_ec2_vpc.secondary_attribute_names == ["vpc_enable_dns_support"]
+@mock_ec2
+@mock_sts
+def test_secondary_attribute_names(ec2_service):
+    assert get_single_ec2_vpc(ec2_service).secondary_attribute_names == ["vpc_enable_dns_support"]
 
 
 @mock_iam
@@ -134,29 +150,27 @@ def test_secondary_attribute_names(single_ec2_vpc):
 def test_secondary_attribute_maps(iam_service):
     create_iam_role()
     single_iam_role = list(iam_service.collection("role").all())[0]
-    assert single_iam_role.get_secondary_attributes_map() == {
-        "InlinePolicyAttachments": {"PolicyNames": ["test-role-policy"], "IsTruncated": False, "Marker": None},
-        "ManagedPolicyAttachments": {
-            "AttachedPolicies": [
-                {
-                    "PolicyName": "APIGatewayServiceRolePolicy",
-                    "PolicyArn": "arn:aws:iam::aws:policy/aws-service-role/APIGatewayServiceRolePolicy",
-                }
-            ],
-            "IsTruncated": False,
-            "Marker": None,
-        },
+    single_iam_role.fetch_secondary_attributes()
+    assert single_iam_role.secondary_attributes_map == {
+        "ManagedPolicyAttachments": [
+            {
+                "PolicyName": "APIGatewayServiceRolePolicy",
+                "PolicyArn": "arn:aws:iam::aws:policy/aws-service-role/APIGatewayServiceRolePolicy",
+            }
+        ],
     }
 
 
-def test_shape(single_ec2_vpc):
-    assert single_ec2_vpc.shape.name == "Vpc"
+@mock_ec2
+@mock_sts
+def test_shape(ec2_service):
+    assert get_single_ec2_vpc(ec2_service).shape.name == "Vpc"
 
 
 @mock_ec2
+@mock_sts
 def test_relationships(ec2_service):
-    single_ec2_vpc = list(ec2_service.collection("vpc").all())[0]
-    assert single_ec2_vpc.relationships[0].partial_urn.resource_type == "dhcp_options"
+    assert get_single_ec2_vpc(ec2_service).relationships[0].partial_urn.resource_type == "dhcp_options"
 
 
 def test_empty_resource(ec2_service):